✨ Shift-attacking monsters now target monster instead of tile
The previous behavior would target the XY ground tile below the monster even when the cursor was selecting the monster when the action was taken.

This improves tracking of the monster (when monsters are moving for example) and allows for detection of targeting when a monster dies, so that the queued attack is cancelled on monster death.